rename plugin.js to plugins.js
